Line num and error handling
Have linenum functionality and error handling for strings. I need to add
in tokenization for spaces inside of a string, and EOF warning, but
otherwise I think I'm pretty much done with lex. Line num works
relatively, so it starts counting lines when it first sees code. I think
this is okay for now, but I may need to fix it.