Id's space char in stings
I hate strings... well, at least dealing with them. I now have
functionality to ID spaces if they are in a string.  The code is a
little messy, so I have to do some clean up and add some clear
documentation as to how it works once the smoke clears.