Test cases and verbose/optimization
Added some test cases that are good for testing my parser (will add more
cases and description to that file too). Further, I'm beginning to
optimize by making one function for each case in matchAndConsume. This
will also work with verbose mode, and put text to the output if the
radio button is selected.