Consistency, very close to multiple program handling
So the power went out, and I now only have an hour of battery left.
Despite this, I'm still going to work this issue: I still haven't gotten
the programming handling. It almost works, but when the src code is
missing a dollar sign where there should be one, it isn't throwing an
error. I'll see what I can do...