CST and some additions. Whoa.
I'm overdue for a commit. In addition to some uncommitted things from
yesterday, which completed the corrections to project 1 except for one
minor detail, I've taken the tree code from the website and started the
CST. I will now be getting into the nitty-grritty of CST, and will
hopefully finish it.