Fixed tab issue. More with CST.
Fixed an issue with tab that the code was having. Now it will process
any program just by copying and pasting in the input (didn't work fully
in the past). Trying to find a way to do one CST per program, but once
that's done, it's on to tackling the AST...