String expressions woooooork
Finally got this to work correctly. Lots of stepping through the code.
The main problem was getting to work consistently with how the grammar
actually works; I was trying to recursively call traverseExpr and not
just call traverseStrExpr on itself (which is consistent with the
grammar). While, If, and different blocks shouldn't be difficult. Have
some test cases to show that my code works.