Hard and annoying pattern changes
Boolean Expr changes when If or While are passed through. This had to be
accounted for in traverseBooleanExpr, and it was annoying as hell to
figure out, but I think it's all fixed up. All I need now is to get it
to handle a different scope, so a new block which is inherent with If or
While, and then I can process multiple ASTs, which will wrap up AST
creation.