Success! Multiple ASTs.
Finally figured it out! I shift the CSTs to get a new workingCST and
create an AST from there, checking to make sure the CST isn't undefined.
This part was all about modeling it like parse, in the sense that I had
to find patterns (in the CSTs), and use recursive decent. Last but not
least, scope and type checking.