Got it
Figured out how to get that while loop to see if it needs to
lookToParentScopes. Type checking is also good to go. I don't see a
situation where these routines won't work, but hey, you never know... I
still have to check IF and WHILE, but then I just have to deal with
warnings, which is much easier.