I've done all I can
This is how I'd summarize what I have so far: it mostly works. There are
some kinks; test cases that don't work. But for the sake of my sanity,
and for the health of those around me, I need to get some sleep. If I
get more time to work on this, awesome. Otherwise, there are other
things I HAVE to get done, and I simply don't have the time to work on
this any more until I get those other tasks accomplished.