Now I've got it. Added some other things.
The problem was that I can't assign a variable to a function call, which
I'm pretty sure you can do in most other programming languages. In any
case, I just needed to remove those checks in my checkBlockChildren;;
the pointer for parent is just fine. Last things are getting if and
while to work properly, some added type checking, and checking for
multiple stages of a longer expression.