One small fix
Not sure if you've looked at this yet professor, but I had a unneeded
check in my checkAssignment function that caused an error (it was due to
me trying to figure out how I would handle a variable being assigned to
another variable). Also, there was a debugger line somewhere, so I
commented that out.