Beginning codeGen, fixed another bug
I had another spot in my code that I didn't return after I found a
semantic error (had to deal with a variable being declared in the same
scope). Starting codeGen as well, just having the basics of displaying
the bytes and soon I'll begin making the prototypes for the jump and
temporary variable table.