Fixed up. One small issue.
Fixed things up so variable types are checked. Only problem I have, is
I'm finding a hard time to get it to distinguish between strings of a
single letter and a variable when I through an error. I can work on this
later if I have the time, but I just want to do some more testing and
then when I'm satisfied with that, we'll see. I think getting into code
gen is much more important. I can maybe leave in the quotes for the
strings and don't display the quotes when I make the ast(s)?