Definitely tested with variety
I now have a well tested project. I made one catch, in that I know test
a long int expr that is printed to make sure that the variable at the
end if there is one is declared. It also ensures that the expression is
valid all together, with or without a variable at the end.