Can print integers!
I can print! INTEGERS. Super cool. Now I just need everything else to
work haha. SvegOS is going to be the go to for testing. This process is
going to look very similar to what semanticAnalysis looks like in terms
of traversing the AST. Just now all I have to do is take whatever is
there and convert it into the corresponding hex code. All the checks are
done at this point, so it's as simple as that...