Can print strings!
Have the functionality to generate strings. This is big, since I now
have the heap dealt with... mostly. I'm not entirely sure that we need
to allocate a certain amount of space for the heap. That's a question
then: what's a good way to detect a stack overflow? It would have to be
when we start overwriting bytes, and that can be done with in operations
of the code itself, the static space, or the heap. For now, I'm just
going to continue. This will be big for going to the next steps. I'll
finish up print, then move on to the others (if, while, assignment and
var decl).