Wow. I can do some crazy stuff.
I can print variables, handle variables, scopes... wow. It's all coming
pretty easy knock on wood. I've been cross testing the code gen with one
of the previous projects in the hall of fame, and I've tested it in
SvegOS... sooooo cool. Kind of stressful though to be honest hahaha,
just because I think something is going to go wrong, but it doesn't! And
it all works! I have to handle longer int expressions still (with
printing and variables),assigning variables to each other, and then I
have to handle if/while... I think I'll be done over the weekend!