More progress
Can assign variables to variables. What we talked about in class about
strings, and just updating the pointer later in the code, is what I
already do. Now that the final is over, I am very ready to be done with
this class and this project! Decided I'll get if and while working
first, and then do longer int expressions for assignment and print.
Then... well, that's it!