A small inconsistency. More complex conditions.
There's a small problem that I did not catch in my AST development; it
the first part of the condition is a string, then it does not produce
the string. Not an issue, since we don't support comparing strings
anyways, but still something to improve for later versions of this
code... haha. I have more complex conditions working, but I still need
to have it work for comparing a variable to a number or a boolean. Then,
there's an issue I have to fix with nested if/while, and then longer
ints for printing or for assignment. Thennnnnn, that's ittttt!