Conditions work
Conditions work... on one condition. The left hand side has to be a
var. This is fine since otherwise there would be an error, and it just
makes more sense for it to work that way programmatically. The only
thing now is figuring out my nesting problem and then just having longer
integers to work.